修改前三个Demo中写错的类型

+++ LearnOpenGLES/Rectangle/OpenGLES_Rectangle.m
@@ -123,7 +123,7 @@ static const float Colors[] = {
     glVertexAttribPointer(_positionSlot, 3, GL_FLOAT, GL_FALSE, 0, Vertices);
     glVertexAttribPointer(_colorSlot,    4, GL_FLOAT, GL_FALSE, 0, Colors);
     
-    glDrawArrays(GL_TRIANGLE_STRIP, 0, sizeof(Vertices) / (sizeof(int) * 3));
+    glDrawArrays(GL_TRIANGLE_STRIP, 0, sizeof(Vertices) / (sizeof(float) * 3));
     
     //把缓冲区的数据呈现到UIView上
     [_context presentRenderbuffer:GL_RENDERBUFFER];
